refactor gc worker
