+++ src/server/gc_worker/mod.rs
@@ -4,6 +4,7 @@ mod applied_lock_collector;
 mod compaction_filter;
 mod config;
 mod gc_manager;
+mod gc_runner;
 mod gc_worker;
 
 // TODO: Use separated error type for GCWorker instead.
@@ -12,7 +13,8 @@ pub use compaction_filter::WriteCompactionFilterFactory;
 use compaction_filter::{is_compaction_filter_allowd, CompactionFilterInitializer};
 pub use config::{GcConfig, GcWorkerConfigManager, DEFAULT_GC_BATCH_KEYS};
 pub use gc_manager::AutoGcConfig;
-pub use gc_worker::{sync_gc, GcSafePointProvider, GcTask, GcWorker, GC_MAX_EXECUTING_TASKS};
+pub use gc_runner::GcTask;
+pub use gc_worker::{GcController, GcSafePointProvider, GC_MAX_EXECUTING_TASKS};
 
 #[cfg(test)]
 pub use compaction_filter::tests::gc_by_compact;
